backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-play-scraper==1.2.5
google-generativeai==0.5.3
httpx==0.27.0
python-dotenv==1.0.1
